fix clippy for all-targets
