+++ .github/workflows/ci.yml
@@ -39,4 +39,4 @@ jobs:
           toolchain: 1.92.0
           components: clippy
       - uses: Swatinem/rust-cache@v2
-      - run: cargo clippy --workspace --all-targets
+      - run: cargo clippy --workspace --all-targets -- -D warnings
